refactor(app): dissolve the last import cycles — src/app is now cycle-free
- den.ts no longer doubles as the Den type registry: the 14 wire types
  consumed across module boundaries (DenSettings, DenUser, DenOrg*,
  DenPluginConfig*, DenResourceSnapshot*) move to new leaf den-types.ts;
  den.ts re-exports them so its many importers are untouched.
  den-session-events and openwork-server now import the leaf instead of the
  2k-line client implementation.
- ReloadReason ownership flips from types.ts to extensions.ts (it is part of
  the extension manifest contract); types.ts re-exports. extensions.ts is
  now a leaf module.
- types.ts imports OpencodeConfigFile/WorkspaceInfo from desktop-types
  (leaf) instead of the desktop bridge implementation.

madge --circular on src/app + src/i18n: 0 cycles (7 when this branch
started; the graph is now a DAG).

Tested:
- pnpm typecheck (clean)
- bun test tests/ — 58 pass, 0 fail
- pnpm build (vite production build clean)
- live: vite dev server 200, den-types module serves
